refactor: Currency enum 클래스 common 패키지로 이동
